requirements.txt
google-generativeai
python-dotenv
langchain
langchain-community
langchain_groq
langchain-huggingface
langchain-google-genai
PyPDF2
faiss-cpu
langchain_google_genai
uvicorn
fastapi
chromadb
pydantic
python-multipart
langchain
openai
pypdf